+++ src/ui/GlobalSettings.cpp
@@ -1,0 +1,3 @@
+#include "GlobalSettings.h"
+
+char GlobalSettings::defaultFileDialogDirectory[512]{ "." };