+++ src/ui/GlobalSettings.h
@@ -1,0 +1,5 @@
+#pragma once
+
+struct GlobalSettings {
+    static char defaultFileDialogDirectory[512];
+};